requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
diff-match-patch==20230430
Django==5.0.7
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-filter==24.3
django-import-export==4.1.1
django-js-asset==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
inflection==0.5.1
Markdown==3.7
msgpack==1.0.8
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
service-identity==24.1.0
setuptools==74.0.0
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
XlsxWriter==3.2.0
zope.interface==7.0.3
